Rename avroUtils into avro_utils
