+++ bin/save_distribution_schema.py
@@ -23,7 +23,7 @@ import json
 
 from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession, load_parquet_files
-from fink_broker.avroUtils import readschemafromavrofile
+from fink_broker.avro_utils import readschemafromavrofile
 from fink_broker.loggingUtils import get_fink_logger, inspect_application
